Patch by Murray Jensen, 20 Jun 2003: - hymod update - cleanup (especially for gcc-3.x compilers)

+++ include/linux/mtd/nftl.h
@@ -90,7 +90,7 @@ struct NFTLrecord {
 	__u16 lastEUN;                  /* should be suppressed */
 	__u16 numfreeEUNs;
 	__u16 LastFreeEUN; 		/* To speed up finding a free EUN */
-	__u32 long nr_sects;
+	__u32 nr_sects;
 	int head,sect,cyl;
 	__u16 *EUNtable; 		/* [numvunits]: First EUN for each virtual unit  */
 	__u16 *ReplUnitTable; 		/* [numEUNs]: ReplUnitNumber for each */
